+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Print season calendar
 - DEV to title in development mode
 - Short str method in championship model
+- Next and previous championships scroll through
+- is_puanla method to championship model
 
 ## [2.2.1] - 2024-03-27
 ### Changed